+++ CHANGELOG.md
@@ -5,7 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.6.1] - 2024-07-31
+
+### Fixed
+
+- `DebugFiltered.LogException` only logs when `filterLevel` is exactly `Exception`
+
+### Development
+
+- Added `DebugFiltered.LogException` tests
+
+
 ## [1.6.0] - 2024-06-06
+
+### Addded
 
 - Add copy constructor to `WeightedRandomList`